+++ CHANGELOG.md
@@ -14,6 +14,13 @@ All notable changes to this project will be documented in this file.
     - Signature::name()
     - Symbol::name()
     - Symbol::string()
+- add FromSymbol trait
+- update clingo-sys to 0.5.2 
+- add functions to PropagateInit
+    - PropagateInit::add_literal()
+    - PropagateInit::add_weight_constraint()
+    - PropagateInit::add_minimize()
+    - PropagateInit::propagate()
 
 ## v0.6.0 - Jan 3, 2020